fix email exist check not done upon account creation
If I understand correctly, this library assumes email address is unique
at least within a tenant.

An account creation is supposed to be blocked if the same email address
is already used in other accounts. Currently EmailExists check is
performed against "username". This does not become an issue if
EMAILISUSERNAME is enabled. However if it's disabled, there can be
multiple accounts that have the same email address.

This becomes a problem when user runs SendUsernameReminder since a given
email address can be associated with multiple usernames.