Changed the auto property to a backing field for the ClaimCollection property on the RelationalUserAccount, so that we can do a null check in the getter.
This property is a mapped EF property which is only being set by EF when retrieved from the database. When the RelationalUserAccount is created by the idmservice and the DbContext is not renewed, EF won't update the Entity resulting in a null exception when doing more operations on the account (like adding claims).