Empty MultiTenant unit test fix to catch empty tenants
In MultiTenant mode. The CreateAccount method allows empty tenants so it
is never required. The Init method should always have a tenant or a
default tenant and if it doesn't then it should throw a TenantRequired
validation exception.

The Unit test for this has been updated to reflect this functionality.