fix(AuthenticationService): SignInWithLinkedAccount sets UserAccount properties too late
Properties on the UserAccount set via the MapClaimsToAccount should occur
before the AccountCreatedEvent fires. Doing so enables flexibility in building
the content of the email sent in response to this event.

Also as a side-effect, there is a slight performance benefit as it removes one
call to SaveChanges on the underlying DbContext